% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{add_noise_contract}
\alias{add_noise_contract}
\title{Add contracted uniform noise to points on sphere.}
\usage{
add_noise_contract(m, contr = NULL, sd = NULL)
}
\arguments{
\item{m}{nx3. points on sohere in R3 coordinates.}

\item{contr}{double in \link{0,1}. Contraction parameter.}
}
\description{
Add contracted uniform noise to points on sphere.
}
