% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{add_noise_normal}
\alias{add_noise_normal}
\title{Add wrapped normal noise to points on sphere.}
\usage{
add_noise_normal(m, sd)
}
\arguments{
\item{m}{nx3. points on sohere in R3 coordinates.}

\item{sd}{positive double. standard deviation of normal distribution.}
}
\description{
Add wrapped normal noise to points on sphere.
}
