% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{angle2R3_1}
\alias{angle2R3_1}
\title{Convert angle representation to R3.}
\usage{
angle2R3_1(x)
}
\arguments{
\item{x}{vetor length 2}
}
\value{
vector length 3
}
\description{
Convert angle representation to R3.
}
