% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clamp1}
\alias{clamp1}
\title{Clamp vector to interval \link{-1, 1}.}
\usage{
clamp1(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector with values in \link{-1, 1}.
}
\description{
Clamp vector to interval \link{-1, 1}.
}
