% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{convert_a2e}
\alias{convert_a2e}
\title{Convert angle representation to R3}
\usage{
convert_a2e(x)
}
\arguments{
\item{x}{nx2}
}
\value{
nx3
}
\description{
The first angle, theta, is in \link{0, pi} an describes longitude / the third
coordinate. The second angle, phi, is in [0, pi) an describes latitude /
rotation around the third axis.
}
