% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convert_a2e_1}
\alias{convert_a2e_1}
\title{Convert angle representation to R3.}
\usage{
convert_a2e_1(x)
}
\arguments{
\item{x}{vetor length 2}
}
\value{
vector length 3
}
\description{
Convert angle representation to R3.
}
