% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{crop1}
\alias{crop1}
\title{Set values larger than 1 to 1 and smaller than -1 to -1.}
\usage{
crop1(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
the vector cropped to \link{-1,1}
}
\description{
Set values larger than 1 to 1 and smaller than -1 to -1.
}
