% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_sphere_regression.R
\name{estimate_cos_pars}
\alias{estimate_cos_pars}
\title{Estimate the three parameters of the (1D) cosine regression.}
\usage{
estimate_cos_pars(x, dist_yq, lambda_n = 80, lambda_max = 3)
}
\arguments{
\item{x}{A vector of length n. The covariates.}

\item{dist_yq}{A matrix of size n x m, where m is the number of test points
q. The distances form the observations to the test points}
}
\value{
A list of estimated parameters a_q, b_q (as vectors of length m) and
lambda (a scalar).
}
\description{
Estimate the three parameters of the (1D) cosine regression.
}
