% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_sphere_regression.R
\name{estimate_cos_pars_ab}
\alias{estimate_cos_pars_ab}
\title{Estimate the a and b parameters of the (1D) cosine regression.}
\usage{
estimate_cos_pars_ab(x, cos_dist_yq, lambda)
}
\arguments{
\item{x}{A vector of length n. The covariates.}

\item{cos_dist_yq}{A matrix of size n x m, where m is the number of test points
q. The cosine of the distances form the observations to the test points}

\item{lambda}{A scalar. The fixed lambda parameter for cosine regression.}
}
\value{
A 2 x m matrix of a(q,lambda) and b(q,lambda) for all m test points q
and the one given lambda
}
\description{
Estimate the a and b parameters of the (1D) cosine regression.
}
