% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lingeo.R
\name{estimate_lingeo}
\alias{estimate_lingeo}
\title{Use geodesic regression to find the best fitting geodesic}
\usage{
estimate_lingeo(x, y, x_new, max_speed, grid_size = 2)
}
\arguments{
\item{x}{covariate, a vector of length n with values in \link{0, 1}}

\item{y}{observations on the sphere (in R3), a nx3 matrix}

\item{x_new}{double vector, new values of the covariate where to evaluate the
estimated function}

\item{max_speed}{a nonnegative scalar double, a bound on the maximum speed of
the geodesic}
}
\description{
Use geodesic regression to find the best fitting geodesic
}
