% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locfre.R
\name{estimate_locfre}
\alias{estimate_locfre}
\title{Local linear Frechet regression on the sphere with cross validation.}
\usage{
estimate_locfre(
  x,
  y,
  x_new,
  adapt = c("loocv", "none"),
  bw = 7,
  kernel = "epanechnikov",
  grid_size = 2
)
}
\arguments{
\item{bw}{the bandwidth $h$ or number of bandwidths to check (for adapt="loocv")}
}
\description{
Uses leave one out cross validation to find a suitable bandwidth
}
