% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locgeo.R
\name{estimate_locgeo}
\alias{estimate_locgeo}
\title{Local geodesic regression on the sphere with cross validation.}
\usage{
estimate_locgeo(
  x,
  y,
  x_new,
  adapt = c("loocv", "none"),
  bw = 7,
  kernel = "epanechnikov",
  max_speed = 10,
  grid_size = 2,
  accuracy = 0.25
)
}
\arguments{
\item{bw}{number of bandwidths to check}

\item{accuracy}{double in \link{0,1}. trade optimization accuracy for
computational speed}
}
\description{
Uses leave one out cross validation to find a suitable bandwidth
}
