% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_v_rcpp}
\alias{get_v_rcpp}
\title{Converts a 2D tangentspace vector representation to 3D}
\usage{
get_v_rcpp(x, a)
}
\arguments{
\item{x}{vetor length 3}

\item{a}{vetor length 2}
}
\value{
vector length 3
}
\description{
Converts a 2D tangentspace vector representation to 3D
}
