% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{r_sphere_contract}
\alias{r_sphere_contract}
\title{Sample from the contracted uniform distribution on the sphere.}
\usage{
r_sphere_contract(n, t, m_a = NULL, m = NULL)
}
\arguments{
\item{n}{number of samples}

\item{t}{double in \link{0,1}. contraction parameter. t=0 means constant m, t=1 is uniform distribution}

\item{m_a}{double vector of length 2. Point to contract to.}
}
\description{
Sample from the contracted uniform distribution on the sphere.
}
