% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{r_sphere_unif}
\alias{r_sphere_unif}
\title{Sample from the uniform distribution on the sphere.}
\usage{
r_sphere_unif(n)
}
\arguments{
\item{n}{number of samples}
}
\description{
Sample from the uniform distribution on the sphere.
}
