% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{rotation_matrix}
\alias{rotation_matrix}
\title{Rotation matrix.}
\usage{
rotation_matrix(m_a)
}
\arguments{
\item{m_a}{Vector of length two. The two angles: theta and phi.}
}
\value{
3x3
}
\description{
A matrix that rotates (0,0,1) to m_a.
Note that there a many such matrices.
This matrix is the combination of a rotation around y- and z-axis.
}
