% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{row_norm3}
\alias{row_norm3}
\title{Calculate the Euclidean norm of an nx3 matrix rowwise.}
\usage{
row_norm3(x)
}
\arguments{
\item{x}{nx3 matrix}
}
\value{
vector of length n
}
\description{
Calculate the Euclidean norm of an nx3 matrix rowwise.
}
