% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{sample_data}
\alias{sample_data}
\title{Sample regression data from a geodesic model with normal noise.}
\usage{
sample_data(n, sd, x_new = NULL, speed_bounds = NULL, p = NULL, v = NULL)
}
\arguments{
\item{n}{positive integer. number of observations to sample.}

\item{sd}{positive double. standard deviation of normal noise.}

\item{x_new}{double vector, new values of the covariate where to evaluate the
estimated function}

\item{speed_bounds}{double vector of length 1 or 2}
}
\value{
a list with observations at x in y and true function values at x_new
in m
}
\description{
Sample regression data from a geodesic model with normal noise.
}
