% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_parallel}
\alias{simulate_parallel}
\title{Simulate using package parallel}
\usage{
simulate_parallel(opt_list, cores = parallel::detectCores(), seed = NULL)
}
\description{
Repetitions of each setting a run in parallel.
}
